Update bytesize_test.go
int will overflow during the 1TB conversion test, leading to a compilation error on some platforms. 

Fix by forcing a uint64 use

+++ afind/bytesize_test.go
@@ -7,7 +7,7 @@ import (
 )
 
 func TestByteSize(t *testing.T) {
-	check := func(nb int, exp string) {
+	check := func(nb uint64, exp string) {
 		bs := ByteSize(nb)
 		if bs.String() != exp {
 			t.Errorf("%v bytes: got %v, want %v", nb, bs.String(), exp)
